updated newUser react component adjusting timezone to float and its c…

+++ client/src/utils/mutations.js
@@ -2,7 +2,7 @@ import { gql } from '@apollo/client';
 
 
 export const ADD_USER = gql`
-    mutation addUser($username: String!, $email: String!, $zodiacName: String!, $timezone: Int!, $password: String!) {
+    mutation addUser($username: String!, $email: String!, $zodiacName: String!, $timezone: Float!, $password: String!) {
         addUser(username: $username, email: $email, zodiacName: $zodiacName, timezone: $timezone, password: $password) {
             token
             user {
